[REFACT] Normalized integer type

+++ scanners/code_scanner.cpp
@@ -235,7 +235,7 @@ size_t pesieve::CodeScanner::collectExecutableSections(RemoteModuleData &_remote
 {
 	size_t initial_size = sections.size();
 	const size_t sec_count = peconv::get_sections_count(_remoteModData.headerBuffer, _remoteModData.getHeaderSize());
-	for (size_t i = 0; i < sec_count; i++) {
+	for (DWORD i = 0; i < sec_count; i++) {
 		PIMAGE_SECTION_HEADER section_hdr = peconv::get_section_hdr(_remoteModData.headerBuffer, _remoteModData.getHeaderSize(), i);
 		if (section_hdr == nullptr) {
 			continue;
